Disable geo coding on update.

+++ docroot/modules/custom/towerhealth_msow_migration/src/EventSubscriber/DoctorPostMigrationSubscriber.php
@@ -56,52 +56,33 @@ class DoctorPostMigrationSubscriber implements EventSubscriberInterface {
    */
   public function onPreImport(MigrateImportEvent $event) {
     $migration = $event->getMigration();
-    $id_map = $migration->getIdMap();
-    $id_map->prepareUpdate();
-    $source = $migration->getSourcePlugin();
-    $source->rewind();
-    $source_id_values = [];
-    while ($source->valid()) {
-      $source_id_values[] = $source->current()->getSourceIdValues();
-      $source->next();
-    }
-    $id_map->rewind();
-    $destination = $migration->getDestinationPlugin();
-    while ($id_map->valid()) {
-      $map_source_id = $id_map->currentSource();
-      if (!in_array($map_source_id, $source_id_values, TRUE)) {
-        $destination_ids = $id_map->currentDestination();
-        $node = Node::load($destination_ids['nid']);
-
-        if ($node instanceof EntityInterface) {
-          $node->set('moderation_state', 'archived');
-          $node->save();
-          //$destination->rollback($destination_ids);
-          $id_map->delete($map_source_id);
-        }
+    if ($migration->id() === 'towerhealth_providers' || $migration->id() === 'towerhealth_locations') {
+      $id_map = $migration->getIdMap();
+      $id_map->prepareUpdate();
+      $source = $migration->getSourcePlugin();
+      $source->rewind();
+      $source_id_values = [];
+      while ($source->valid()) {
+        $source_id_values[] = $source->current()->getSourceIdValues();
+        $source->next();
       }
-      /**
-       * /*$this->dispatchRowDeleteEvent(MigrateEvents::PRE_ROW_DELETE, $migration, $destination_ids);
-      $this->dispatchRowDeleteEvent(MigratePlusEvents::MISSING_SOURCE_ITEM, $migration, $destination_ids);
-      $this->dispatchRowDeleteEvent(MigrateEvents::POST_ROW_DELETE, $migration, $destination_ids);
-       */
-      $id_map->next();
+      $id_map->rewind();
+      $destination = $migration->getDestinationPlugin();
+      while ($id_map->valid()) {
+        $map_source_id = $id_map->currentSource();
+        if (!in_array($map_source_id, $source_id_values, TRUE)) {
+          $destination_ids = $id_map->currentDestination();
+          $node = Node::load($destination_ids['nid']);
+          // Don't remove the node from the system, only archive it.
+          if ($node instanceof EntityInterface) {
+            $node->set('moderation_state', 'archived');
+            $node->save();
+            $id_map->delete($map_source_id);
+          }
+        }
+        $id_map->next();
+      }
+      $this->dispatcher->dispatch(MigrateEvents::POST_ROLLBACK, new MigrateRollbackEvent($migration));
     }
-    $this->dispatcher->dispatch(MigrateEvents::POST_ROLLBACK, new MigrateRollbackEvent($migration));
-  }
-
-  /**
-   * Dispatches MigrateRowDeleteEvent event.
-   *
-   * @param string $event_name
-   *   The event name to dispatch.
-   * @param \Drupal\migrate\Plugin\MigrationInterface $migration
-   *   The active migration.
-   * @param array $destination_ids
-   *   The destination identifier values of the record.
-   */
-  protected function dispatchRowDeleteEvent($event_name, MigrationInterface $migration, array $destination_ids) {
-    // Symfony changing dispatcher so implementation could change.
-    //$this->dispatcher->dispatch($event_name, new MigrateRowDeleteEvent($migration, $destination_ids));
   }
 }
